:recycle: Moved summarization stuff to summarize
