+++ src/summarize/summarize/__init__.py
@@ -5,6 +5,26 @@ from sumy.nlp.tokenizers import Tokenizer
 from sumy.summarizers.lsa import LsaSummarizer as Summarizer
 from sumy.nlp.stemmers import Stemmer
 from sumy.utils import get_stop_words
+from bs4 import BeautifulSoup
+import re
+
+
+def extract_useful(
+    soup: BeautifulSoup,
+    headers: Tuple[str, ...] = ("classroom activities", "turned in", "deliverables"),
+) -> str:
+    """Extracts 'Classroom Activities', 'Deliverables', etc"""
+    # TODO: Deprecate and use JS parser
+    name_re = re.compile("|".join(f"({x})" for x in headers))
+
+    def query(tag) -> bool:
+        return (
+            tag.name == "h4"
+            and tag.strong
+            and name_re.search(tag.strong.get_text().lower()) is not None
+        )
+
+    return "".join(map(lambda x: str(x) + str(x.next_sibling or ""), soup(query)))
 
 
 def algorithm(html, language: str = "english", sentence_count: int = 10):
